requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements/prod.in
#
aiohttp==3.8.1
    # via gradio
aiosignal==1.2.0
    # via aiohttp
analytics-python==1.4.0
    # via gradio
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
asgiref==3.5.2
    # via uvicorn
async-timeout==4.0.2
    # via aiohttp
asynctest==0.13.0
    # via aiohttp
attrs==20.3.0
    # via aiohttp
backoff==1.10.0
    # via analytics-python
bcrypt==3.2.2
    # via paramiko
boto3==1.24.40
    # via smart-open
botocore==1.27.40
    # via
    #   boto3
    #   s3transfer
cached-property==1.5.2
    # via h5py
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.2
    # via uvicorn
coloredlogs==15.0.1
    # via onnxruntime
cryptography==37.0.2
    # via paramiko
cycler==0.11.0
    # via matplotlib
et-xmlfile==1.1.0
    # via openpyxl
fastapi==0.78.0
    # via gradio
ffmpy==0.3.0
    # via gradio
filelock==3.8.0
    # via
    #   huggingface-hub
    #   transformers
flatbuffers==22.9.24
    # via onnxruntime
fonttools==4.33.3
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.5.0
    # via gradio
gradio==3.0.21
    # via -r requirements/prod.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
h5py==3.6.0
    # via -r requirements/prod.in
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via gradio
huggingface-hub==0.10.0
    # via
    #   timm
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
importlib-metadata==4.11.3
    # via
    #   -r requirements/prod.in
    #   click
    #   huggingface-hub
    #   transformers
intel-openmp==2022.2.0
    # via mkl
jinja2==2.11.3
    # via
    #   -r requirements/prod.in
    #   gradio
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kiwisolver==1.4.2
    # via matplotlib
linkify-it-py==1.0.3
    # via markdown-it-py
markdown-it-py[linkify,plugins]==2.1.0
    # via
    #   gradio
    #   mdit-py-plugins
markupsafe==1.1.1
    # via
    #   -r requirements/prod.in
    #   jinja2
matplotlib==3.5.2
    # via gradio
mdit-py-plugins==0.3.0
    # via markdown-it-py
mdurl==0.1.1
    # via markdown-it-py
mkl==2022.2.0
    # via mkl-service
mkl-service==2.4.0
    # via -r requirements/prod.in
monotonic==1.6
    # via analytics-python
mpmath==1.2.1
    # via sympy
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.21.6
    # via
    #   -r requirements/prod.in
    #   gradio
    #   h5py
    #   matplotlib
    #   onnxruntime
    #   openai
    #   pandas
    #   torchvision
    #   transformers
onnxruntime==1.12.1
    # via -r requirements/prod.in
openai==0.23.1
    # via -r requirements/prod.in
openpyxl==3.0.10
    # via openai
orjson==3.7.2
    # via gradio
packaging==20.9
    # via
    #   huggingface-hub
    #   matplotlib
    #   onnxruntime
    #   transformers
pandas==1.3.5
    # via
    #   gradio
    #   openai
pandas-stubs==1.2.0.62
    # via openai
paramiko==2.11.0
    # via gradio
pillow==7.1.2
    # via
    #   -r requirements/prod.in
    #   gradio
    #   matplotlib
    #   torchvision
protobuf==3.20.3
    # via onnxruntime
pycparser==2.21
    # via cffi
pycryptodome==3.14.1
    # via gradio
pydantic==1.9.1
    # via
    #   fastapi
    #   gradio
pydub==0.25.1
    # via gradio
pynacl==1.5.0
    # via paramiko
pyngrok==5.1.0
    # via -r requirements/prod.in
pyparsing==2.4.2
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   analytics-python
    #   botocore
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via -r requirements/prod.in
python-multipart==0.0.5
    # via gradio
pytz==2022.1
    # via pandas
pyyaml==5.4.1
    # via
    #   huggingface-hub
    #   pyngrok
    #   timm
    #   transformers
regex==2022.9.13
    # via transformers
requests==2.27.1
    # via
    #   -r requirements/prod.in
    #   analytics-python
    #   gradio
    #   huggingface-hub
    #   openai
    #   torchvision
    #   transformers
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   analytics-python
    #   mkl-service
    #   paramiko
    #   python-dateutil
    #   python-multipart
smart-open[s3]==5.2.1
    # via -r requirements/prod.in
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.19.1
    # via fastapi
sympy==1.10.1
    # via onnxruntime
tbb==2021.7.0
    # via mkl
timm==0.6.11
    # via -r requirements/prod.in
tokenizers==0.12.1
    # via transformers
torch==1.12.0
    # via
    #   -r requirements/prod.in
    #   timm
    #   torchvision
torchvision==0.13.0
    # via
    #   -r requirements/prod.in
    #   timm
tqdm==4.64.0
    # via
    #   -r requirements/prod.in
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.22.2
    # via -r requirements/prod.in
typing-extensions==4.1.1
    # via
    #   aiohttp
    #   anyio
    #   asgiref
    #   async-timeout
    #   huggingface-hub
    #   importlib-metadata
    #   kiwisolver
    #   markdown-it-py
    #   openai
    #   pandas-stubs
    #   pydantic
    #   starlette
    #   torch
    #   torchvision
    #   uvicorn
    #   yarl
uc-micro-py==1.0.1
    # via linkify-it-py
urllib3==1.26.12
    # via
    #   botocore
    #   requests
uvicorn==0.17.6
    # via gradio
yarl==1.7.2
    # via aiohttp
zipp==3.8.0
    # via importlib-metadata
